refactor: improve types and move chain to stack
